pyproject.toml
[project]
name = "agenticore"
version = "0.1.0"
description = "Claude Code runner and orchestrator — thin job lifecycle, repo management, and OTEL pipeline"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "The Cloud Clock Work"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0",
    "redis>=7.0",
    "uvicorn[standard]>=0.30",
    "httpx>=0.25",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/The-Cloud-Clock-Work/agenticore"
Repository = "https://github.com/The-Cloud-Clock-Work/agenticore"
Issues = "https://github.com/The-Cloud-Clock-Work/agenticore/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "fakeredis>=2.21",
    "ruff>=0.4",
]

[project.scripts]
agenticore = "agenticore.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agenticore*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: no external deps",
    "integration: requires Docker",
]

[tool.ruff]
target-version = "py312"
line-length = 120
